# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SystemGatewayAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Modules": {
      "post": {
        "tags": [
          "Modules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Modules"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Modules/{ModuleId}": {
      "put": {
        "tags": [
          "Modules"
        ],
        "parameters": [
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Modules"
        ],
        "parameters": [
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Modules"
        ],
        "parameters": [
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Modules/{ModuleId}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Modules"
        ],
        "parameters": [
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Patient/{Token}/Modules/{ModuleId}": {
      "post": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Patient/{Token}/Modules/{ModuleId}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Patient/{token}/Modules/{ModuleId}": {
      "get": {
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Therapist/{token}/Patients/{email}/Modules/{ModuleId}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{Token}/Patients/{PatientEmail}/Modules/{ModuleId}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{Token}/Patients/{PatientEmail}/Modules/{ModuleId}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataPoint": {
        "type": "object",
        "properties": {
          "sectionName": {
            "type": "string",
            "nullable": true
          },
          "isDataEditable": {
            "type": "boolean"
          },
          "content": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleData": {
            "$ref": "#/components/schemas/ModuleVersion"
          }
        },
        "additionalProperties": false
      },
...
      "ModuleVersion": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "nullable": true
          },
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "dataStructure": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPoint"
            },
            "nullable": true
          },
          "htmlCard": {
            "type": "string",
            "nullable": true
          },
          "htmlDashboard": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checksum": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}